feature: sort button deleted and view setting button modified
