--- src/platform/assets/components/MediaAssetSortButton.vue
@@ -1,19 +1,0 @@
-<template>
-  <div class="inline-flex items-center">
-    <Popover>
-      <template #button>
-        <Button variant="secondary" size="icon">
-          <i class="icon-[lucide--arrow-up-down]" />
-        </Button>
-      </template>
-      <template #default="{ close }">
-        <slot :close />
-      </template>
-    </Popover>
-  </div>
-</template>
-
-<script setup lang="ts">
-import Button from '@/components/ui/button/Button.vue'
-import Popover from '@/components/ui/Popover.vue'
-</script>
